Initialize the samplers when creating Bird
The code spends a lot of time initializaing the samplers during the
random walk. Since all data are available when the Bird is initialized,
we initialize the samplers then and store them in a slice. We gain in
performance at the expense of a bigger memory footprint.

For 2,000,000 items & 1,000,000 users & a query of size 10,000 we
currently have an output time of 5ms.