Validate the input  to NewBird
It is best not to trust the user of the library, and check the validity
of the input instead of letting the program crash later on. It leads to
easier debugging.
The impact on performance should be minimal, as the check is only
performed once every time Bird is initialized.

We also tidied the logs and the doc a bit and renamed a couple of
variables whose name was not clear.