remove log statements
It's not a bad practice to log information from a library, but using
directly the `log` package can be annoying when the application using
the library use something else.

The proper way would be to have an interface for logger objects, and
inject the logger during initialization. In our case, we could simply
use the `log` package from go-service, and accept a `*log.Log` object.

For the time being, the only logs here are timing information, and these
should not be in the library anyway.