fix errors messages
Wrapping errors allows us to provide context to error messages.
Generally speaking, if a function is `DoFoo`, the wrapping message
should explain what failed *in* `DoFoo`, instead of saying "cannot do
foo". The caller of `DoFoo` however, will probably wrap the error
returned by `DoFoo` with the annotation "cannot do foo".

You will notice that the whole "let's annotate and unwind the stack" is
called "throwing an exception" is a lot of other languages. But I guess
it's a bit too modern for Golang, and doing it manually is so much fun.