Add support for weighted graphs through `Emu`
The relations between users and items are usually weighted, as a users
might interact differently with items. It was currently not poissble
with Bird to weight the relationship.

This commit solves the problem by introducing Emu, which takes a
weighted bipartite graph as an input. It uses the weights to initialize
the samplers associated with each users and returns a Bird.