Add a trust-based recommendation mechanism
We add a new mechanism to recommend items based on trust. We assign to
each referrer a weight that is equal to the number of times this
referrer has been traversed by the random walks. This amounts to
trusting users who have the largest number of tracks in commong more.

Then, for each item traversed by the walk, we sum the respective weights
of their referrer. We then produce recommendations by ordering them by
weight.