restructure the code in `weaver`
Weaver uses different function names to differentiate itself. This is
however unnecessary; while Bird's method are indeed promoted to Weaver,
Weaver's methods override these. We thus rename Weavers' methods to
follow the conventions followed in the rest of the library.

The value of `DefaultWeight` cannot currently be set by the user. We
thus create a `WeaverCfg` struct that contains the parameter, can be set
by hand and passed to the Weaver initialization.

We changed the README accordingly.